requirements.txt
# Core dependencies
numpy>=1.24.0
pillow>=10.0.0
PyOpenGL>=3.1.0
python-osc>=1.8.0
SpoutGL>=0.9.0

# Deep learning (PyTorch)
# Install PyTorch with CUDA support first:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# Then install:
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
accelerate>=0.20.0
